% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bioregion.R
\name{data_bioregion}
\alias{data_bioregion}
\title{Get Bio Region Information Data}
\usage{
data_bioregion(bioregion = "Scotian Shelf")
}
\arguments{
\item{bioregion}{name of the bio region to match from the \code{NAME_E} column from \href{https://gisp.dfo-mpo.gc.ca/arcgis/rest/services/FGP/Federal_Marine_Bioregions/MapServer/0}{Open Data}. Default is \code{"Scotian Shelf"}}
}
\value{
sf data.frame
}
\description{
This function pulls information (length, area, shape, etc.) about
number of bio regions. For a list of available options visit
https://gisp.dfo-mpo.gc.ca/arcgis/rest/services/FGP/Federal_Marine_Bioregions/MapServer/0
}
\examples{
\dontrun{
bioregion <- data_bioregion()
}
}
