% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_commercial_fishing.R
\name{data_commercial_fishing}
\alias{data_commercial_fishing}
\title{Get Eastern Canada Commercial Fishing data}
\usage{
data_commercial_fishing(region)
}
\arguments{
\item{region}{sf data.frame of the bio region. You can use \code{MarConsNetData::data_planning_area()} to download the planning area of your choice.}
}
\value{
a "sf" "dataframe" object
}
\description{
This function gets the "All Species" dataset of species/gear type commercial fisheries from 2012 to 2021 in the Eastern Canada Regions in the specified bio region. For more
information visit
https://open.canada.ca/data/dataset/502da2ef-bffa-4d9b-9e9c-a7425ff3c594
}
\examples{
\dontrun{
planning <- data_planning_areas()
fishing <- data_commercial_fishing(planning)
}

}
