% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_context.R
\name{data_context}
\alias{data_context}
\title{Get Context from Canadian Protected and Conserved Areas}
\usage{
data_context(type = NULL, area = "St. Anns Bank Marine Protected Area")
}
\arguments{
\item{type}{argument of either \code{site} or \code{network} indicating which objectives to obtain}

\item{area}{a name of an area of which to obtain the objectives from. Options include
\code{st_Anns_Bank_MPA}, \code{musquash_MPA}, \code{laurentian_Channel_MPA}, \code{gully_MPA}, \code{gilbert_MPA}, \code{eastport_MPA},
\code{basin_MPA}, \code{bancsDesAmericains_MPA}, \code{WEBCA}}
}
\value{
a "sf" "dataframe" object
}
\description{
This function gets site and network level context for the
Canadian protected and conserved areas in the Scotian Shelf bio region.
}
\examples{
\dontrun{
stAnnsBankContext <- data_context(type="site", area="st_Anns_Bank_MPA")
}

}
