% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_CPCAD_areas.R
\name{data_CPCAD_areas}
\alias{data_CPCAD_areas}
\title{Get Canadian Protected and Conserved Areas Database}
\usage{
data_CPCAD_areas(bioregion, zones = FALSE)
}
\arguments{
\item{bioregion}{sf data.frame of the bio region. You can use \code{MarConsNetData::get_bioregion()} to download the bio region of your choice.}

\item{zones}{logical that indicates if zones should be returned (\code{TRUE}) or, as default, unioned (\code{FALSE}). If zones is \code{FALSE}, the sf data.frame is grouped by \code{NAME_E} and the geoms are unioned resulting in a complete loss of other columns.}
}
\value{
a "sf" "dataframe" object
}
\description{
This function gets polygons of the the Canadian protected
and conserved areas in the specified bio region. For more
information visit
https://www.canada.ca/en/environment-climate-change/services/
national-wildlife-areas/protected-conserved-areas-database.html
}
\examples{
\dontrun{
bioregion <- data_bioregion()
MPAs <- data_CPCAD_areas(bioregion,  zones = FALSE)
}

}
