% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_DFO_sp_lists.R
\name{data_DFO_sp_lists}
\alias{data_DFO_sp_lists}
\title{Extract species lists from DFO website}
\usage{
data_DFO_sp_lists(status = "At-Risk")
}
\arguments{
\item{status}{character string that determines which list is returned. Default is "At-Risk", but "Invasive","","Special Concern",and "all" are also valid options.}
}
\value{
a dataframe
}
\description{
Extract species lists from DFO website
}
\examples{
# get Aquatic Invasive Species list
AIS <- data_DFO_sp_lists(status="Invasive")

# get Species at Risk list
SARA <- data_DFO_sp_lists(status="At-Risk")
}
