% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_draft_areas.R
\name{data_draft_areas}
\alias{data_draft_areas}
\title{Get draft conservation network sites}
\usage{
data_draft_areas(bioregion = "Scotian Shelf")
}
\arguments{
\item{bioregion}{name of the bio region to match from the \code{NAME_E} column from \href{https://gisp.dfo-mpo.gc.ca/arcgis/rest/services/FGP/Federal_Marine_Bioregions/MapServer/0}{Open Data}. Default is \code{"Scotian Shelf"}}
}
\value{
sf data.frame
}
\description{
This function pulls information (length, area, shape, etc.) about
draft conservation network sites for the Scotian Shelf-Bay of Fundy Bioregion. More information is available at:
https://open.canada.ca/data/en/dataset/bb048082-bc05-4588-b4f0-492b1f1b8737
}
\examples{
\dontrun{
draftareas <- data_draft_areas()
}
}
