% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geomorphic_units.R
\name{data_geomorphic_units}
\alias{data_geomorphic_units}
\title{Get Geomorphic Unit Data}
\usage{
data_geomorphic_units()
}
\value{
sf dataframe object
}
\description{
This function gets the geomorphic unit information of the Scotian Shelf.
This is based on the Greenlaw classification system.
}
\examples{
\dontrun{
geomorph <- data_geomorphic_units()
}
}
