% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_indicators.R
\name{data_indicators}
\alias{data_indicators}
\title{Get Monitoring Indicators}
\usage{
data_indicators()
}
\value{
a "sf" "dataframe" object
}
\description{
This function obtains a list of indicators from the monitoring framework for
St. Anns Bank Area of Interest paper and the Ecosystem Based Management paper.
}
\examples{
\dontrun{
stAnnsBankContext <- data_indicators()
}

}
