% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_KBA.R
\name{data_KBA}
\alias{data_KBA}
\title{Get Key Biological Areas (KBA) Data}
\usage{
data_KBA(bioregion, crs = sf::st_crs(bioregion))
}
\arguments{
\item{bioregion}{sf data.frame of the bio region. You can use \code{MarConsNetData::data_bioregion()} to download the bio region of your choice.}

\item{crs}{target coordinate reference system: object of class crs, or input string for st_crs. Default is that of the bioregion.}
}
\value{
"sf" "data.frame" object
}
\description{
Get Key Biological Areas (KBA) Data
}
\examples{
\dontrun{
bioregion <- data_bioregion()
kba <- data_KBA(bioregion)
}
}
