% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_LFA.R
\name{data_LFA}
\alias{data_LFA}
\title{Get Lobster Fishing Areas}
\usage{
data_LFA(bioregion = "Scotian Shelf", layer = 14)
}
\arguments{
\item{bioregion}{name of the bio region to match from the \code{NAME_E} column from \href{https://gisp.dfo-mpo.gc.ca/arcgis/rest/services/FGP/Federal_Marine_Bioregions/MapServer/0}{Open Data}. Default is \code{"Scotian Shelf"}}

\item{layer}{integer of the layer from Open Data to query. Defaults to 14 for the current (as of 2022) boundaries. See the
\href{https://egisp.dfo-mpo.gc.ca/arcgis/rest/services/open_data_donnees_ouvertes/historical_lobster_fishing_districts_maritimes_region/MapServer/}{REST server} for additional options}
}
\value{
sf data.frame
}
\description{
This function pulls information (length, area, shape, etc.) about
Lobster Fishing Areas for the Scotian Shelf-Bay of Fundy Bioregion. More information is available at:
https://ouvert.canada.ca/data/dataset/b8550fde-fbfe-4556-a592-1ae972834b65
}
\examples{
\dontrun{
LFA <- data_LFA()
}
}
