% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_MSRV_survey_open.R
\name{data_MarRV_survey_open}
\alias{data_MarRV_survey_open}
\title{Download the Maritimes Summer Research Vessel Survey data from open.canada.ca}
\usage{
data_MarRV_survey_open(survey = "summer")
}
\arguments{
\item{survey}{character string indicating the survey to be downloaded. Defaults to "summer", but can also accept "spring", "fall", and "4VSW"}
}
\value{
list of data.frames
}
\description{
Data is from the csv files on https://open.canada.ca/data/en/dataset/8ddcaeea-b806-4958-a79f-ba9ab645f53b
}
\examples{
summer <- data_MarRV_survey_open()

fall <- data_MarRV_survey_open(survey="fall")

}
