% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objectives.R
\name{data_objectives}
\alias{data_objectives}
\title{Get Objectives from Canadian Protected and Conserved Areas}
\usage{
data_objectives(
  type = NULL,
  area = "St. Anns Bank Marine Protected Area",
  prohibiton = FALSE
)
}
\arguments{
\item{type}{argument of either \code{site} or \code{network} indicating which objectives to obtain}

\item{area}{a name of an area of which to obtain the objectives from.}

\item{prohibiton}{a boolean indicating if you want the prohibition included}
}
\value{
a "sf" "dataframe" object
}
\description{
This function gets site and network level objectives for the
Canadian protected and conserved areas in the Scotian Shelf bio region.
}
\examples{
\dontrun{
stAnnsBankObjectives <- data_objectives(type="site", area="st_Anns_Bank_MPA")
}

}
