% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_PAVICS.R
\name{data_PAVICS}
\alias{data_PAVICS}
\title{Get climate data}
\usage{
data_PAVICS(
  OPENDAP_url =
    "https://pavics.ouranos.ca/twitcher/ows/proxy/thredds/dodsC/datasets/simulations/RCM-CMIP6/CORDEX/NAM-12/day/NAM-12_NorESM2-MM_historical_r1i1p1f1_OURANOS_CRCM5_v1-r1_day_19500101-20141231.ncml",
  first_date = "1950-01-01",
  end_date = "1951-01-1",
  variable = "tas",
  geom = data_planning_areas()
)
}
\arguments{
\item{OPENDAP_url}{character string for an OPENDAP url from the THREDDS \href{https://pavics.ouranos.ca/thredds/catalog.html}{catalog}. Defaults to "https://pavics.ouranos.ca/twitcher/ows/proxy/thredds/dodsC/datasets/simulations/RCM-CMIP6/CORDEX/NAM-12/day/NAM-12_NorESM2-MM_historical_r1i1p1f1_OURANOS_CRCM5_v1-r1_day_19500101-20141231.ncml"}

\item{first_date}{character string for the first date to be included in the data subset. Defaults to "1950-01-01"}

\item{end_date}{character string for the last date to be included in the data subset. Defaults to "1950-01-01"}

\item{variable}{character string for the variable to query from the data. Defaults to "tas" for Near-Surface Air Temperature. See the OPENDAP url for more options}

\item{geom}{sf data.frame used to limit the spatial extent of the data subset (via bounding box). Defaults to the Scotian Shelf and Bay of Fundy planning area}
}
\value{
matrix
}
\description{
Get data from \href{"https://pavics.ouranos.ca/index.html"}{PAVICS} THREDDS \href{https://pavics.ouranos.ca/thredds/catalog.html}{catalog}
}
\examples{
require(MarConsNetData)
ssbof <- data_planning_areas()
climate <- data_PAVICS
image(apply(climate,c(1,2),mean))

}
