% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_planning_areas.R
\name{data_planning_areas}
\alias{data_planning_areas}
\title{Get Marine Spatial Planning Areas Data}
\usage{
data_planning_areas(PA = "Scotian Shelf and Bay of Fundy")
}
\arguments{
\item{PA}{name of the planning area to match from the \code{NAME_E} column from \href{https://gisp.dfo-mpo.gc.ca/arcgis/rest/services/FGP/Eastern_Canada_Marine_Spatial_Planning_Areas/MapServer/0}{Open Data ESRI REST}. Default is \code{"Scotian Shelf and Bay of Fundy"}}
}
\value{
sf data.frame
}
\description{
This function pulls information (length, area, shape, etc.) about
\href{https://open.canada.ca/data/en/dataset/f089a3f3-45e9-47de-b1c4-170e9950d8e7}{Eastern Canada Marine Spatial Planning Areas}.
}
\examples{
\dontrun{
planning_area <- data_planning_areas()
}
}
