% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SBA.R
\name{data_SBA}
\alias{data_SBA}
\title{Get Coral and Sponge Significant Benthic Areas in Eastern Canada (2016) Data}
\usage{
data_SBA(region)
}
\arguments{
\item{region}{sf data.frame of the region. You can use \code{MarConsNetData::data_planning_areas()} to download the region of your choice.}
}
\value{
a "sf" "dataframe" object
}
\description{
This function gets the Coral and Sponge Significant Benthic Areas
in the specified region. For more information see
https://open.canada.ca/data/en/dataset/6af357a3-3be1-47d5-9d1f-e4f809c4c903
}
\examples{
\dontrun{
region <- data_planning_areas()
sba <- data_SBA(region)
}
}
