% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wind_energy_areas.R
\name{data_wind_energy_areas}
\alias{data_wind_energy_areas}
\title{data_wind_energy_areas
Download and read the Wind Energy Areas (WEA) dataset from the Canada-Nova Scotia Offshore Energy Regulator \href{https://cnsoer.ca/renewable-energy/lands-management/governments-designated-offshore-wind-energy-areas}{(CNSOER)}.}
\usage{
data_wind_energy_areas()
}
\value{
An sf object containing the Wind Energy Areas (WEA) dataset.
}
\description{
data_wind_energy_areas
Download and read the Wind Energy Areas (WEA) dataset from the Canada-Nova Scotia Offshore Energy Regulator \href{https://cnsoer.ca/renewable-energy/lands-management/governments-designated-offshore-wind-energy-areas}{(CNSOER)}.
}
\examples{
\dontrun{
wea <- data_wind_energy_areas()
}
}
