% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biophysical_units.R
\name{get_biophysical_units}
\alias{get_biophysical_units}
\title{get_biophysical_units}
\usage{
get_biophysical_units()
}
\value{
sf dataframe
}
\description{
get_biophysical_units
}
\examples{
biophys <- get_biophysical_units()
}
