% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_COD_url.R
\name{get_COD_url}
\alias{get_COD_url}
\title{Get Canadian Open Data (COD) resource URL by extension}
\usage{
get_COD_url(uuid, ext)
}
\arguments{
\item{uuid}{universally unique identifier of the dataset in the https://open.canada.ca/ portal (e.g., "2d9cce9a-d634-4b49-879f-87c40c52acf2")}

\item{ext}{desired file extension (e.g., "csv", "json", "xlsx")}
}
\value{
A character vector of URLs for resources matching the specified extension.
}
\description{
Get Canadian Open Data (COD) resource URL by extension
}
\examples{
url <- get_COD_url("2d9cce9a-d634-4b49-879f-87c40c52acf2", "csv")
datadict <- read.csv(url)
}
