% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geomorphic_units.R
\name{get_geomorphic_units}
\alias{get_geomorphic_units}
\title{get_geomorphic_units}
\usage{
get_geomorphic_units()
}
\value{
sf dataframe
}
\description{
get_geomorphic_units
}
\examples{
geomorph <- get_geomorphic_units()
}
