% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kba.R
\name{get_kba}
\alias{get_kba}
\title{get_kba}
\usage{
get_kba(bioregion, crs = sf::st_crs(bioregion))
}
\arguments{
\item{bioregion}{sf polygon to constrain query.}

\item{crs}{target coordinate reference system: object of class crs, or input string for st_crs. Default is that of the bioregion.}
}
\description{
get_kba
}
