% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obis_data.R
\name{get_obis_data}
\alias{get_obis_data}
\title{get_obis_data}
\usage{
get_obis_data(areas, geom_col = "geoms", name_col = "NAME_E")
}
\arguments{
\item{areas}{}
}
\value{
data.frame
}
\description{
get_obis_data
}
\examples{
\dontrun{
obis <- testPowerBI::getobisdata()
}
}
