% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_data.R
\name{get_project_data}
\alias{get_project_data}
\title{Get data for the Marine Protected Area App}
\usage{
get_project_data(ids = NULL, titles = NULL, taxize = FALSE, MPA = NULL)
}
\arguments{
\item{ids}{a single or vector of ids of project(s) associated with the Project Planning Tool (PPT)}

\item{titles}{a character indicating the name of a project. See the table in the vignette for more details.}

\item{taxize}{a Boolean indicating if the data frame names should
be categorized by subphylum. If \code{FALSE} the species names will be
given. Note that if the subphylum is "Vertabrae" the class will be
given instead}

\item{MPA}{the name of an MPA in Atlantic Canada. Options include
\code{BancDesAmericains}, \code{basinHead}, \code{eastport}, \code{gilbertBay}, \code{gully},
\code{laurentianChannel}, \code{musquashEstuary}, or \code{stAnnsBank}. If \code{NULL}
all data plotted}
}
\value{
dataframe
}
\description{
This function obtains information required for the
analysis of the Marine Protect Areas (MPAs).
}
\examples{
\dontrun{
data <- get_project_data(id=1093, taxize=FALSE)
}
}
\author{
Jaimie Harbin and Remi Daigle
}
