% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snowCrabSurvey.R
\name{get_snowCrabSurvey}
\alias{get_snowCrabSurvey}
\title{Get data for the snow crab survey (1093)}
\usage{
get_snowCrabSurvey(taxize = FALSE, MPA = NULL)
}
\arguments{
\item{taxize}{a Boolean indicating if the data frame names should
be categorized by subphylum. If \code{FALSE} the species names will be
given. Note that if the subphylum is "Vertabrae" the class will be
given instead}

\item{MPA}{the name of an MPA in Atlantic Canada. Options include
\code{BancDesAmericains}, \code{basinHead}, \code{eastport}, \code{gilbertBay}, \code{gully},
\code{laurentianChannel}, \code{musquashEstuary}, or \code{stAnnsBank}. If \code{NULL}
all data plotted}
}
\value{
dataframe
}
\description{
This function obtains the species counts identified with the snow crab survey
}
\examples{
\dontrun{
data <- get_snowCrabSurvey(taxize=TRUE)
}
}
