% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getobisdata.R
\name{getobisdata}
\alias{getobisdata}
\title{getobisdata}
\usage{
getobisdata(bioregion = "Scotian Shelf")
}
\arguments{
\item{bioregion}{name of the bioregion to match from the `NAME_E` column of: https://gisp.dfo-mpo.gc.ca/arcgis/rest/services/FGP/Federal_Marine_Bioregions/MapServer/0}
}
\value{
data.frame
}
\description{
getobisdata
}
\examples{
\dontrun{
obis <- testPowerBI::getobisdata()
}
}
