% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{package}
\name{MarConsNetData-package}
\alias{MarConsNetData}
\alias{MarConsNetData-package}
\title{A Package for Locating and Transforming MPA Data}
\description{
The \code{MarConsNetData} package is to locate and transform
real data collected in a variety of different Marine Protected Areas (MPAs).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dfo-mar-mpas/MarConsNetData}
  \item \url{https://dfo-mar-mpas.github.io/MarConsNetData/}
  \item Report bugs at \url{https://github.com/dfo-mar-mpas/MarConsNetData/issues}
}

}
