% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_AZMP.R
\name{project_AZMP}
\alias{project_AZMP}
\title{Get data for AZMP (579)}
\usage{
project_AZMP()
}
\value{
dataframe
}
\description{
This function uses the azmpdata package to get the location of relevant
sampling used
}
