% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_optimizingeDNA.R
\name{project_optimizingeDNA}
\alias{project_optimizingeDNA}
\title{Get data for the optimizing eDNA survey (1491)}
\usage{
project_optimizingeDNA(taxize = FALSE, MPA = NULL)
}
\arguments{
\item{taxize}{a Boolean indicating if the data frame names should
be categorized by subphylum. If \code{FALSE} the species names will be
given. Note that if the subphylum is "Vertabrae" the class will be
given instead}

\item{MPA}{the name of an MPA in Atlantic Canada. Options include
\code{BancDesAmericains}, \code{basinHead}, \code{eastport}, \code{gilbertBay}, \code{gully},
\code{laurentianChannel}, \code{musquashEstuary}, or \code{stAnnsBank}. If \code{NULL}
all data plotted}
}
\description{
This function obtains the species counts identified with the
optimizing eDNA survey.
}
