% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_whale_biodiversity.R
\name{project_whale_biodiversity}
\alias{project_whale_biodiversity}
\title{Obtain Whale Sighting data}
\usage{
project_whale_biodiversity()
}
\value{
a data frame containing number of whale sightings per year for WEBCA
}
\description{
Obtain Whale Sighting data
}
