% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize_data.R
\name{taxize_data}
\alias{taxize_data}
\title{Group species into taxonomic groups}
\usage{
taxize_data(df = NULL)
}
\arguments{
\item{df}{data frame to be taxized}
}
\description{
This function groups species into taxonomic groups. It is used in the \code{taxize=TRUE}
argument of the \code{get_project_data} function.
}
