% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_biophysical_units.R
\name{data_biophysical_units}
\alias{data_biophysical_units}
\title{Get Biophysical Units Data}
\usage{
data_biophysical_units()
}
\value{
sf dataframe
}
\description{
This function obtains information (lenght, area, etc.) of smaller regions
in the Scotian shelf that are classified using the "Greenlaw classification"
system. For more information visit Open Data at
https://open.canada.ca/data/en/dataset/abbfe250-8202-4822-8bd1-aaaf393bbfd5
}
\examples{
\dontrun{
biophys <- data_biophysical_units()
}
}
