% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SAR_CH.R
\name{data_SAR_CH}
\alias{data_SAR_CH}
\title{Get Species at Risk Critical Habitat Data}
\usage{
data_SAR_CH(bioregion)
}
\arguments{
\item{bioregion}{sf data.frame of the bio region. You can use \code{MarConsNetData::get_bioregion()} to download the bio region of your choice.}
}
\value{
"sf" "data.frame" object
}
\description{
Get Species at Risk Critical Habitat Data
}
\examples{
\dontrun{
bioregion <- data_bioregion()
SARCH <- data_SAR_CH(bioregion)
}
}
