% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CPCAD_areas.R
\name{get_CPCAD_areas}
\alias{get_CPCAD_areas}
\title{Title}
\usage{
get_CPCAD_areas(bioregion, zones = FALSE)
}
\arguments{
\item{bioregion}{sf data.frame of the bioregion. You can use \code{MarConsNetData::get_bioregion()} to download the bioregion of your choice.}

\item{zones}{logical that indicates if zones should be returned (\code{TRUE}) or, as default, unioned (\code{FALSE}). If zones is \code{FALSE}, the sf data.frame is grouped by \code{NAME_E} and the geoms are unioned resulting in a complete loss of other columns.}
}
\description{
Title
}
\examples{
bioregion <- get_bioregion()
MPAs <- get_CPCAD_areas(bioregions,  zones = FALSE)

}
