% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{_PACKAGE}
\name{MarConsNetData-package}
\alias{MarConsNetData-package}
\title{A Package for Locating and Transforming MPA Data}
\description{
The \code{MarConsNetData} package is to locate and transform
real data collected in a variety of different Marine Protected Areas (MPAs).
}
