% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_AZMP.R
\name{project_AZMP}
\alias{project_AZMP}
\title{Get data for AZMP (579)}
\usage{
project_AZMP(bd = bloom_df)
}
\arguments{
\item{bd}{bloom_df likely from the targets file}
}
\value{
dataframe
}
\description{
This function uses the azmpdata package to get the location of relevant
sampling used
}
