% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_AZMP.R
\name{project_AZMP}
\alias{project_AZMP}
\title{Get data for AZMP (579)}
\usage{
project_AZMP(season = "fall", destdir = "tmp", force = FALSE)
}
\arguments{
\item{season}{character vector of either fall or spring. If fall is given
the data from the fall program is returned. If spring is given the
CTD data from the spring program is returned. If summer or winter
is given, the CTD ground fish data is given from the respective
season.}

\item{destdir}{a directory indicating where to save the fetched data.}

\item{force}{A boolean indicating if a new version of this data should
be obtained or not}
}
\value{
A list of NetCDF files downloaded in your destdir
}
\description{
This function gets the CTD data from the fall and spring
surveys as well as the fixed stations and ground fish CTD
data.
}
